src/gcs_to_bq/requirements.txt
functions-framework==3.8.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0

